UI for Fish Tracker
Shows which fish have been caught and which fish have not been caught.

Created Fish Tracker UI script which sets up labels and assigns fish to them for each fish in the available fish list on the fish tracker.

Created Fish Label UI script which assigns the name of the text based on the fish they are set with as well as the color of the text to red to indicate they haven't been caught yet.

When a fish is caught if its the fish assigned to the label the label will change the text color to green to show the fish has been caught.